refactor: #385: split the `helpers` file to the transaction entity
