+++ src/entities/position/index.ts
@@ -1,2 +1,3 @@
 export { Positions } from './ui/table';
 export type { PositionsTableProps } from './ui/table';
+export { updatePositionsQuery } from './api/use-positions';
